refactor avatar image upload with tanstack
